pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pythia"
version = "0.1.0"
description = "A codebase indexing and inquiry system using LEANN for semantic search with MCP interface"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Pythia Contributors" }
]
keywords = [
    "mcp",
    "semantic-search",
    "codebase-indexing",
    "leann",
    "rag",
    "vector-database",
    "git",
    "kubernetes"
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    "leann>=0.3.0",
    "mcp>=1.0.0",
    "openai>=1.0.0",
    "gitpython>=3.1.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "httpx>=0.25.0",
    "click>=8.0.0",
    "rich>=13.0.0",
    "watchdog>=3.0.0",
    "aiofiles>=23.0.0",
    "tenacity>=8.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
    "pre-commit>=3.0.0",
]
kubernetes = [
    "kubernetes>=28.0.0",
]

[project.scripts]
pythia = "pythia.cli:main"
pythia-mcp = "pythia.mcp_server:main"
pythia-agent = "pythia.agent:main"

[project.urls]
Homepage = "https://github.com/panbanda/pythia"
Documentation = "https://github.com/panbanda/pythia#readme"
Repository = "https://github.com/panbanda/pythia"
Issues = "https://github.com/panbanda/pythia/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/pythia"]

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by formatter)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["pythia"]

[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "-v --cov=pythia --cov-report=term-missing"

[tool.coverage.run]
source = ["src/pythia"]
branch = true

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
